requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.59.2
imageio==2.37.0
importlib_resources==6.5.2
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.4
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
tifffile==2024.8.30
tzdata==2025.2
zipp==3.23.0
